Sparse Matrix Class implemented
Matrix multiplication and tensor products working as expected.
Added adding magic method for both matrix classes to ensure that the dimensions of the matrices match. This is done to prevent incorrect adding of the np.ndarray objects.
Added time and scipy packages to test performance, shows that for 16x16 sparse matrices of 10% density, Sparse class performs about 26x better.