Fixes Sparse Class
Elements attribute split into elements and indices.
[elements, i , j] -> [elements] [i, j]
This way we can take advantage of numpy arrays without having the issue of indices being integers while elements are complex. I.e. numpy is not designed to handle mixed data types, dtype can be set to object but not designed for that.