Redesigned the CNOT gate function, added extra functionality by allowing multi-target and multi-control gates
Redesigned the CNOT gate function from scratch. Now CNOT function takes in a list of control qubits and a list of target qubits and constructs an n-qubit CNOT gate.
This allows for CNOT-like gates such as the toffoli to be created from this function alone. For example for the toffoli gate, we input 2 control qubits and 1 target qubit to create the gate.